RDKTV-35484 Add deviceIdentification plugin details into deviceinfo (#6190)
* RDKTV-35484 Add deviceIdentification plugin details into deviceinfo

Reason for change: Add chipset property to DeviceInfo Plugin
Test Procedure: verify build success and basic test
Risks: Low
Priority: P1

* RDKTV-35484 Add deviceIdentification plugin details into deviceinfo

Reason for change: Add chipset property to DeviceInfo Plugin
Test Procedure: verify build success and basic test
Risks: Low
Priority: P1

---------
