--- DeviceIdentification/CHANGELOG.md
@@ -1,53 +1,0 @@
-# Changelog
-
-All notable changes to this RDK Service will be documented in this file.
-
-* Each RDK Service has a CHANGELOG file that contains all changes done so far. When version is updated, add a entry in the CHANGELOG.md at the top with user friendly information on what was changed with the new version. Please don't mention JIRA tickets in CHANGELOG. 
-
-* Please Add entry in the CHANGELOG for each version change and indicate the type of change with these labels:
-    * **Added** for new features.
-    * **Changed** for changes in existing functionality.
-    * **Deprecated** for soon-to-be removed features.
-    * **Removed** for now removed features.
-    * **Fixed** for any bug fixes.
-    * **Security** in case of vulnerabilities.
-
-* Changes in CHANGELOG should be updated when commits are added to the main or release branches. There should be one CHANGELOG entry per JIRA Ticket. This is not enforced on sprint branches since there could be multiple changes for the same JIRA ticket during development. 
-
-* For more details, refer to [versioning](https://github.com/rdkcentral/rdkservices#versioning) section under Main README.
-
-## [1.0.6] - 2024-12-30
-### Fixed
-- Added method for retrieving Serial number if existing implementation fails to retrive the serial number.
-
-## [1.0.6] - 2024-05-28
-### Fixed
-- Added methods for retrieving Serial number, Chip Id and Firmware version for Broadcom Devices.
-
-## [1.0.5] - 2024-01-10
-### Fixed
-- Add chipset name in DeviceIdentification plugin
-
-## [1.0.4] - 2023-09-12
-### Added
-- Implement Thunder Plugin Configuration for Kirkstone builds(CMake-3.20 & above)
-
-## [1.0.3] - 2023-06-05
-### Added 
-- Added Support to build the plugin to both R4 & R2
-
-## [1.0.2] - 2022-11-03
-### Fixed
-- Fixed DeviceIdentification plugin activation issue
-
-## [1.0.1] - 2022-10-04
-### Fixed
-- Fixed warnings that are treated as errors with "-Wall -Werror"
-
-## [1.0.0] - 2022-05-11
-### Added
-- Add CHANGELOG
-
-### Change
-- Reset API version to 1.0.0
-- Change README to inform how to update changelog and API version
