--- DeviceIdentification/DeviceIdentification.config
@@ -1,22 +1,0 @@
-set(autostart ${PLUGIN_DEVICEIDENTIFICATION_AUTOSTART})
-set(preconditions Platform)
-
-
-if(PLUGIN_DEVICEIDENTIFICATION_STARTUPORDER)
-set(startuporder ${PLUGIN_DEVICEIDENTIFICATION_STARTUPORDER})
-endif()
-
-map()
-if(GENERIC_DEVICEIDENTIFIATION)
-    if(PLUGIN_DEVICEIDENTIFICATION_INTERFACE_NAME)
-        kv(interface ${PLUGIN_DEVICEIDENTIFICATION_INTERFACE_NAME})
-    endif()
-endif()
-
-    key(root)
-    map()
-        kv(mode ${PLUGIN_DEVICEIDENTIFICATION_MODE})
-    end()
-end()
-
-ans(configuration)
