--- DeviceIdentification/DeviceIdentificationPlugin.json
@@ -1,13 +1,0 @@
-{
-    "$schema": "https://raw.githubusercontent.com/rdkcentral/rdkservices/main/Tools/json_generator/schemas/plugin.schema.json",
-    "info": {
-        "title": "DeviceIdentification Plugin",
-        "callsign": "DeviceIdentification",
-        "locator": "libWPEFrameworkDeviceIdentification.so",
-        "status": "production",
-        "description": "The `DeviceIdentification` plugin allows you to retrieve various device-related information"
-    },
-    "interface": {
-        "$ref": "DeviceIdentification.json#"
-    }
-}
