+++ DeviceInfo/DeviceInfo.conf.in
@@ -3,6 +3,10 @@ startuporder = "@PLUGIN_DEVICEINFO_STARTUPORDER@"
 
 configuration = JSON()
 
+if boolean("@GENERIC_DEVICEIDENTIFIATION@"):
+    if boolean("@PLUGIN_DEVICEIDENTIFICATION_INTERFACE_NAME@"):
+        configuration.add("interface", @PLUGIN_DEVICEIDENTIFICATION_INTERFACE_NAME@)
+
 rootobject = JSON()
 rootobject.add("mode", "@PLUGIN_DEVICEINFO_MODE@")
 configuration.add("root", rootobject)
