+++ DeviceInfo/DeviceInfo.config
@@ -7,6 +7,12 @@ set (startuporder ${PLUGIN_DEVICEINFO_STARTUPORDER})
 endif()
 
 map()
+if(GENERIC_DEVICEIDENTIFIATION)
+    if(PLUGIN_DEVICEIDENTIFICATION_INTERFACE_NAME)
+        kv(interface ${PLUGIN_DEVICEIDENTIFICATION_INTERFACE_NAME})
+    endif()
+endif()
+
     key(root)
     map()
         kv(mode ${PLUGIN_DEVICEINFO_MODE})
