+++ DeviceInfo/Implementation/MFR/MFR.cpp
@@ -18,6 +18,7 @@
  */
 
 #include "../../Module.h"
+#include <interfaces/IDeviceIdentification.h>
 
 #include <fstream>
 
@@ -27,7 +28,7 @@ extern "C" {
 
 namespace WPEFramework {
 namespace Plugin {
-    class DeviceImplementation : public PluginHost::ISubSystem::IIdentifier {
+    class DeviceImplementation : public Exchange::IDeviceIdentification, public PluginHost::ISubSystem::IIdentifier {
 
     public:
         DeviceImplementation()
@@ -70,8 +71,28 @@ namespace Plugin {
             return _firmwareVersion;
         }
 
+        // IDeviceIdentification interface
+
+        Core::hresult Identification(DeviceInfo& info) const override
+        {
+            info.deviceID = "";
+            uint8_t myBuffer[64];
+
+            myBuffer[0] = Identifier(sizeof(myBuffer) - 1, &(myBuffer[1]));
+
+            if (myBuffer[0] != 0) {
+                info.deviceID = Core::SystemInfo::Instance().Id(myBuffer, ~0);
+            }
+
+            info.firmwareVersion = FirmwareVersion();
+            info.chipset = Chipset();
+
+            return Core::ERROR_NONE;
+        }
+
         BEGIN_INTERFACE_MAP(DeviceImplementation)
         INTERFACE_ENTRY(PluginHost::ISubSystem::IIdentifier)
+        INTERFACE_ENTRY(Exchange::IDeviceIdentification)
         END_INTERFACE_MAP
 
     private:
