+++ DeviceInfo/cmake/FindMFRFWLibs.cmake
@@ -6,6 +6,12 @@
 #
 # Copyright 2020 RDK Management.
 #
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
 # are met:
