+++ DeviceInfo/cmake/FindNXCLIENT.cmake
