+++ Tests/L1Tests/CMakeLists.txt
@@ -61,7 +61,6 @@ set_source_files_properties(
 
 include_directories(../../LocationSync
         ../../SecurityAgent
-        ../../DeviceIdentification
         ../../DeviceDiagnostics
         ../../FrameRate
         ../../AVInput
@@ -102,7 +101,6 @@ include_directories(../../LocationSync
         )
 link_directories(../../LocationSync
         ../../SecurityAgent
-        ../../DeviceIdentification
         ../../DeviceDiagnostics
         ../../FrameRate
         ../../AVInput
@@ -144,7 +142,6 @@ target_link_libraries(${PROJECT_NAME}
         ${NAMESPACE}DeviceDiagnostics
         ${NAMESPACE}LocationSync
         ${NAMESPACE}SecurityAgent
-        ${NAMESPACE}DeviceIdentification
         ${NAMESPACE}FrameRate
         ${NAMESPACE}AVInput
         ${NAMESPACE}ScreenCapture
