+++ l1tests.cmake
@@ -186,7 +186,6 @@ set(PLUGIN_DEVICEDIAGNOSTICS ON)
 set(PLUGIN_LOCATIONSYNC ON)
 set(PLUGIN_TIMER ON)
 set(PLUGIN_SECURITYAGENT ON)
-set(PLUGIN_DEVICEIDENTIFICATION ON)
 set(PLUGIN_FRAMERATE ON)
 set(PLUGIN_AVINPUT ON)
 set(PLUGIN_TELEMETRY OFF)
